membuat crud pembayaran

+++ application/controllers/Admin.php
@@ -8,7 +8,7 @@ class Admin extends CI_Controller {
 		$this->load->model('m_model');
 		$this->load->helper('my_helper');
 		$this->load->library('upload');
-        if($this->session->userdata('logged_in')!=true){
+		if($this->session->userdata('logged_in')!=true && $this->session->userdata('role') != 'admin'){
             redirect(base_url().'auth');
         }
 	}
@@ -19,7 +19,7 @@ class Admin extends CI_Controller {
 			$this->load->view('admin/index', $data);
 		}
 
-	public function upload_img($value)
+	public function upload_image_siswa($value)
 		{
 			$kode = round(microtime(true) * 1000);
 			$config['upload_path'] = './images/siswa/';	
@@ -35,6 +35,24 @@ class Admin extends CI_Controller {
 				return array(true, $nama);
 			}
 		}
+
+	public function upload_image_admin($value)
+		{
+			$kode = round(microtime(true) * 1000);
+			$config['upload_path'] = './images/admin/';
+			$config['allowed_types'] = 'jpg|png|jpeg';
+			$config['max_size'] = 30000;
+			$config['file_name'] = $kode;
+			$this->upload->initialize($config);
+			if (!$this->upload->do_upload($value)) {
+				return array(false, '');
+			} else {
+				$fn = $this->upload->data();
+				$nama = $fn['file_name'];
+				return array(true, $nama);
+			}
+		}
+	
 	
 	public function siswa()
 		{
@@ -51,7 +69,7 @@ class Admin extends CI_Controller {
 		
 	public function aksi_tambah_siswa()
 		{
-			$foto = $this->upload_img('foto');
+			$foto = $this->upload_image_siswa('foto');
 			if ($foto[0] == false) {
 				$data = [
 					'foto' => 'User.png',
@@ -77,10 +95,41 @@ class Admin extends CI_Controller {
 			
 		}
 
+	// public function hapus_siswa($id)
+	// 	{
+	// 		$this->m_model->delete('siswa', 'id_siswa', $id);
+	// 		redirect(base_url('admin/siswa'));
+	// 	}
+
 	public function hapus_siswa($id)
 		{
-			$this->m_model->delete('siswa', 'id_siswa', $id);
-			redirect(base_url('admin/siswa'));
+			$siswa = $this->m_model->get_by_id('siswa', 'id_siswa', $id)->row();
+			if($siswa){
+				if($siswa->foto !== 'User.png'){
+					$file_path = './images/siswa/' . $siswa->foto;
+
+					if(file_exists($file_path)){
+						if(unlink($file_path)){
+							//hapus file berhasil menggunakan model delete
+							$this->m_model->delete('siswa', 'id_siswa',$id);
+							redirect(base_url('admin/siswa'));
+						}else{
+							//gagal menghapus file
+							echo "gagal menghapus file.";
+						}
+					}else{
+						//file tidak di temukan
+						echo "File Tidak Di Temukan"; 
+					}
+				}else{
+					//Tanpa Hapus File User.png
+					$this->m_model->delete('siswa', 'id_siswa', $id);
+					redirect(base_url('admin/siswa'));
+				}
+			}else{
+				//Siswa Tidak Di Temukan
+				echo "Siswa Tidak Di Temukan";
+			}
 		}
 
 	public function ubah_siswa($id)
@@ -117,43 +166,61 @@ class Admin extends CI_Controller {
 		}
 
 	public function aksi_ubah_akun()
-    {
-        $password_baru = $this->input->post('password_baru');
-        $konfirmasi_password = $this->input->post('konfirmasi_password');
-        $email = $this->input->post('email');
-        $username = $this->input->post('username');
+		{
+			$foto = $this->upload_image_admin('foto');
+			if ($foto[0] == false) {
+				$password_baru = $this->input->post('password_baru');
+				$konfirmasi_password = $this->input->post('konfirmasi_password');
+				$email = $this->input->post('email');
+				$username = $this->input->post('username');
+				$data = [
+					 'foto' => 'User.png',
+					'email' => $email,
+					'username' => $username,
+				];
+				if (!empty($password_baru)) {
+					if ($password_baru === $konfirmasi_password) {
+						$data['password'] = md5($password_baru);
+					} else {
+						$this->session->set_flashdata('message', 'Password baru dan Konfirmasi password harus sama');
+						redirect(base_url('admin/akun'));
+					}
+				}
+				$this->session->set_userdata($data);
+				$update_result = $this->m_model->update('admin', $data, array('id' => $this->session->userdata('id')));
+	
+				if ($update_result) {
+					redirect(base_url('admin/akun'));
+				} else {
+					redirect(base_url('admin/akun'));
+				}
+			} else {
+				$password_baru = $this->input->post('password_baru');
+				$konfirmasi_password = $this->input->post('konfirmasi_password');
+				$email = $this->input->post('email');
+				$username = $this->input->post('username');
+				$data = [
+					 'foto' => $foto[1],
+					'email' => $email,
+					'username' => $username,
+				];
+				if (!empty($password_baru)) {
+					if ($password_baru === $konfirmasi_password) {
+						$data['password'] = md5($password_baru);
+					} else {
+						$this->session->set_flashdata('message', 'Password baru dan Konfirmasi password harus sama');
+						redirect(base_url('admin/akun'));
+					}
+				}
+				$this->session->set_userdata($data);
+				$update_result = $this->m_model->update('admin', $data, array('id' => $this->session->userdata('id')));
+	
+				if ($update_result) {
+					redirect(base_url('admin/akun'));
+				} else {
+					redirect(base_url('admin/akun'));
+				}
+			}
+		}
 
-        // Buat data yang akan diubah
-        $data = [
-            'email' => $email,
-            'username' => $username,
-        ];
-
-        // Jika ada passwrod baru
-        if (!empty($password_baru)) {
-            // Pastikan password baru dan konfirmasi password sama
-            if ($password_baru === $konfirmasi_password) {
-                // Hash password baru
-                $data['password'] = md5($password_baru);
-            } else {
-                $this->session->set_flashdata(
-                    'message',
-                    'Password baru dan konfirmasi password harus sama'
-                );
-                redirect(base_url('admin/akun'));
-            }
-        }
-
-        // Lakukan pembaruan data
-        $this->session->set_userdata($data);
-        $update_result = $this->m_model->ubah_data('admin', $data, [
-            'id' => $this->session->userdata('id'),
-        ]);
-
-        if ($update_result) {
-            redirect(base_url('admin/akun'));
-        } else {
-            redirect(base_url('admin/akun'));
-        }
-    }
 	}
